Move Vision class to the vision package
