+++ src/main/java/frc/robot/subsystems/drivetrain/Drivetrain.java
@@ -27,9 +27,9 @@ import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine.Config;
 import frc.robot.Constants.DrivetrainConstants;
 import frc.robot.Constants.VisionConstants;
+import frc.robot.subsystems.vision.Vision;
 import frc.robot.Constants.LoggingConstants;
 import frc.robot.util.Util;
-import frc.robot.subsystems.Vision;
 
 import static edu.wpi.first.units.Units.MetersPerSecond;
 import static edu.wpi.first.units.Units.RadiansPerSecond;
