+++ src/main/java/frc/robot/subsystems/vision/Vision.java
@@ -1,8 +1,7 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.vision;
 
 import edu.wpi.first.math.VecBuilder;
 import frc.robot.Constants.VisionConstants;
-import frc.robot.subsystems.vision.LimelightHelpers;
 import swervelib.SwerveDrive;
 
 /**
